Divide tests in different files. Remove unnecesary files and comments.
