--- framework/src/models/index.ts
@@ -1,4 +1,0 @@
-export * from "./Order";
-export * from "./Inventory";
-export * from "./responses/Response";
-export * from "./responses/DeleteOrderResponse";
